# docs/swagger.yaml
definitions:
  models.Transaction:
    properties:
      customer:
        type: string
      id:
        type: integer
      price:
        type: number
      quantity:
        type: integer
      request_id:
        type: integer
      timestamp:
        type: string
    type: object
  models.TransactionRequest:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Transaction'
        type: array
      request_id:
        type: integer
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
  /dummy-transactions/{dummy}:
    get:
      description: Creating a new dummy Transaction.
      parameters:
      - description: Total of Dummy Transaction required
        in: path
        name: dummy
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TransactionRequest'
      summary: Create Dummy Transaction.
      tags:
      - Transaction
  /save-transactions:
    post:
      description: Creating a new Transaction.
      parameters:
      - description: the body to create a new Transaction
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
      summary: Create New Transaction.
      tags:
      - Transaction
  /transactions:
    get:
      description: Get a list of Transaction.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Transaction'
            type: array
      summary: Get all Transaction.
      tags:
      - Transaction
swagger: "2.0"